requirements.txt
streamlit
sqlalchemy
pymysql
langchain
langchain-community
requests
python-dotenv
